Migrate translateString from function to component
